Added more scraps from previous project
Threw in a handful of the old classes from ProjectFlounder.

The only files actually somewhat implemented are:
- tile.cpp
- tileMap.cpp

The others were just edited so that no errors exist, but most likely
would require some additional work to implement.

Almost all of this needs to be cut down and re-written. This just
provides a nice place to start.